DjangoRestAPI/requirements.txt
alembic==0.8.10
argcomplete==1.9.3
attrs==17.4.0
autopep8==1.3.5
backports-abc==0.5
base58==1.0.0
beautifulsoup4==4.6.3
bleach==2.1.2
boto==2.49.0
boto3==1.9.53
botocore==1.12.53
cairocffi==0.9.0
CairoSVG==2.1.3
certifi==2018.11.29
cffi==1.11.5
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
cmake==3.12.0
configparser==3.5.0
croniter==0.3.24
cssselect2==0.2.1
cycler==0.10.0
DateTime==4.2
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
dill==0.2.8.2
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.30.0
django-anymail==0.7
django-debug-toolbar==1.6
django-dotenv==1.4.2
django-easy-pdf==0.1.1
django-extensions==2.1.9
django-filter==2.1.0
django-heroku==0.3.1
django-import-export==1.0.1
django-model-utils==2.6
django-rest-auth==0.9.0
django-storages==1.6.6
django-webpack-loader==0.4.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
dlib==19.15.0
docopt==0.6.2
docutils==0.14
durationpy==0.5
elasticsearch==6.3.1
enum-compat==0.0.2
enum34==1.1.6
et-xmlfile==1.0.1
flake8==3.5.0
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.10
gunicorn==19.9.0
hjson==3.0.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
ipython-genutils==0.2.0
itsdangerous==0.24
jdcal==1.4
Jinja2==2.8.1
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
kappa==0.6.0
kiwisolver==1.0.1
lambda-packages==0.20.0
lockfile==0.12.2
luigi==2.7.5
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
nbformat==4.4.0
numpy==1.14.5
oauthlib==2.1.0
odfpy==1.3.6
opencv-python==3.4.2.17
openpyxl==2.5.5
ordereddict==1.1
pandas==0.23.1
pdfrw==0.4
Pillow==5.2.0
placebo==0.8.2
plotly==3.0.0
psutil==4.4.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
py4j==0.10.7
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.4
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
pyspark==2.3.1
python-daemon==2.1.2
python-dateutil==2.6.1
python-decouple==3.0
python-editor==1.0.3
python-nvd3==0.14.2
python-openid==2.2.5
python-slugify==1.2.4
python3-openid==3.1.0
python-dotenv
pytz==2018.7
PyYAML==3.12
redis==2.10.6
reportlab==3.5.4
requests==2.20.1
requests-oauthlib==1.0.0
retrying==1.3.3
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
setproctitle==1.1.10
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smmap2==2.0.3
SQLAlchemy==1.2.9
sqlparse==0.2.4
tablib==0.12.1
tabulate==0.7.7
thrift==0.11.0
tinycss2==0.6.1
toml==0.10.0
tornado==4.5.3
tqdm==4.19.1
traitlets==4.3.2
troposphere==2.3.3
typing==3.6.4
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.24.1
virtualenv==16.0.0
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
wsgi-request-logger==0.4.6
WTForms==2.2.1
xhtml2pdf==0.2.2
xlrd==1.1.0
xlwt==1.3.0
zope.deprecation==4.3.0
zope.interface==4.5.0
